refactor(dev-workflow/feed-platform-ms-01-workspace-foundation/T-E+T-H+T-I/r2): split smoke tests into feature/*.test.ts colocated with verification targets
Step 7 → Step 8 transition での user-gate-driven structural refinement (path
restructure / ENV detection と同 user gate 内の追加要望) として、test ファイル
配置を **検証対象 module の隣に colocation** する形に変更する。User 要望:
「smoke.test.ts は全て検証ロジックの近くに .test.ts を作成。feature/*.test.ts
とするのが適切」。

旧設計はソース直下 `smoke.test.ts` 1 ファイルに全検証を集約していたが、
test 内容と検証対象 module の対応が直感的に追いづらく、検証対象が増えた際の
拡張点も曖昧だった。新設計では検証対象ごとに専用 test ファイルに分割し、
検証対象の隣に置く co-location 形に改める。

変更内容 (実装ファイル — 既に Main 側で適用済):

- 削除 (3 件):
  - js/app/feed-platform-backend/src/smoke.test.ts
  - js/app/feed-platform-web/app/smoke.test.ts
  - js/app/identity-provider/app/smoke.test.ts
- 新規追加 (5 件):
  - js/app/feed-platform-backend/src/feature/health.test.ts (Health 1 件、
    describe 'Health'、relative import ./env.ts ./health.ts)
  - js/app/feed-platform-web/app/feature/greeting.test.ts (Greeting 1 件、
    describe 'Greeting'、relative import ./greeting.ts)
  - js/app/feed-platform-web/app/feature/health.test.ts (Health 1 件、
    describe 'Health'、relative import ./env.ts ./health.ts)
  - js/app/identity-provider/app/feature/greeting.test.ts (web と同形)
  - js/app/identity-provider/app/feature/health.test.ts (web と同形)
- 合計 5 件 (test 件数 / 内容 / PASS-FAIL は完全不変、配置と describe ラベル
  と import path のみ変更)

Artifact 更新 (本 commit の主目的):

- design.md:
  - Project A / B / C のディレクトリツリー: smoke.test.ts → feature/*.test.ts
  - A-7 / B-7 / C-7 snippet: 配置パス + 内容を実際の colocation 形に書き換え
  - Test placement policy 説明 (CC-5 周辺) を colocation 規約に更新
  - SC-3 mapping 表 / Performance expectations / Handoff notes (T-E / T-H /
    T-I) を新形式に
  - 既存 Path restructure deviation note 内の "src/smoke.test.ts は不変"
    記述を新形式に補足
  - 末尾に "Deviation note: Test placement switched to feature/<name>.test.ts
    colocation (Step 7 → Step 8 user-gate refinement, 2026-05-06)" を追加
    (経緯 / 変更内容 / 設計上の利点 / SC への影響 / re_activations counter /
    実装完了 commit)

- qa-design.md:
  - Test file placement policy セクションを colocation 規約に更新
  - TC-004 Pass criterion / Notes を 5 件 colocation test PASS 観測に拡張
  - その他 smoke 言及 (Rationale / SC-3 mapping / 統計) を新表現に統一

- task-plan.md:
  - T-E / T-H / T-I の Summary / Artifact / Test cases / Design references
    を新形式に
  - T-G の mermaid label 更新
  - 末尾に "Test placement deviation note (Step 7 → Step 8 user-gate-driven)"
    セクション新設

- TODO.md:
  - T-E / T-H / T-I notes に test placement refinement の deviation 記録を
    追加 (re_activations は加算しない、structural-only)

- ADR-01:
  - Newly added セクションのファイル一覧で smoke.test.ts → feature/*.test.ts
    に更新
  - Existing impact に "Step 7 → Step 8 user-gate refinement (test placement)"
    項目を追加
  - Constraints going forward に "テストファイルは検証対象 module の隣に
    colocation" 規約を ms-02 以降への継承事項として追記

検証 (Main 側で確認済、本実装者再確認済):

- vp check exit 1 (rss-graphql 既存 85 errors 許容、本サイクル雛形内 0)
- vp test run js/app exit 0 (Test Files 17 passed / Tests 83 passed)
- find js/app -name '*.test.ts' で本サイクル雛形は 5 ファイル確認 (1 backend
  + 2 web + 2 IdP)

re_activations counter: 本変更は Step 7 review-driven Blocker / Major findings
ではなく Step 7 → Step 8 transition での user-gate-driven structural refinement
のため、TODO.md の T-E / T-H / T-I の re_activations は加算しない (notes 欄に
deviation 記録のみ)。
