--- js/app/feed-platform-web/app/smoke.test.ts
@@ -1,27 +1,0 @@
-import { Effect, Layer } from 'effect'
-import { describe, expect, test } from 'vite-plus/test'
-
-import * as Env from '#@/feature/env.ts'
-import * as Greeting from '#@/feature/greeting.ts'
-import * as Health from '#@/feature/health.ts'
-
-describe('feed-platform-web smoke', () => {
-  test('Greeting.greet returns expected string via Layer.provide + Effect.runPromise', async () => {
-    const program = Effect.gen(function* () {
-      const greeting = yield* Greeting.Service
-      return greeting.greet('feed-platform-web')
-    })
-    const result = await Effect.runPromise(Effect.provide(program, Greeting.layer))
-    expect(result).toBe('Hello, feed-platform-web')
-  })
-
-  test('Health.check returns ok with env injected from layer', async () => {
-    const program = Effect.gen(function* () {
-      const checker = yield* Health.Service
-      return yield* checker.check()
-    })
-    const layer = Health.layer.pipe(Layer.provide(Env.makeLayer({ ENV: 'development' })))
-    const result = await Effect.runPromise(Effect.provide(program, layer))
-    expect(result).toEqual({ env: 'development', status: 'ok' })
-  })
-})
