+++ js/app/identity-provider/app/feature/health.test.ts
@@ -1,0 +1,17 @@
+import { Effect, Layer } from 'effect'
+import { describe, expect, test } from 'vite-plus/test'
+
+import * as Env from './env.ts'
+import * as Health from './health.ts'
+
+describe('Health', () => {
+  test('check returns ok with env injected from layer', async () => {
+    const program = Effect.gen(function* () {
+      const checker = yield* Health.Service
+      return yield* checker.check()
+    })
+    const layer = Health.layer.pipe(Layer.provide(Env.makeLayer({ ENV: 'development' })))
+    const result = await Effect.runPromise(Effect.provide(program, layer))
+    expect(result).toEqual({ env: 'development', status: 'ok' })
+  })
+})
